refactor telegram for mcp
